feat: Integrate FlutterFire for Firebase services
This commit adds the 'com.google.gms.google-services' plugin to the `android/app/build.gradle` file. This plugin is essential for integrating FlutterFire, enabling the application to utilize Firebase services such as authentication, database, and cloud messaging.  The addition of this plugin allows the app to connect to and use Firebase.